requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altgraph==0.17.4
asgiref==3.8.1
astroid==2.4.2
asyncio==3.4.3
attrs==24.2.0
black==24.10.0
build==1.2.2.post1
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
disutils==1.4.32.post2
Django==5.1.3
django-cors-headers==4.5.0
django-debug-toolbar==4.4.6
django-filebrowser-no-grappelli==4.0.2
django-tinymce==4.1.0
djangorestframework==3.15.2
djlint==1.36.1
EditorConfig==0.12.4
frozenlist==1.4.1
idna==3.10
isort==5.5.2
jsbeautifier==1.15.1
jsmin==3.0.1
json5==0.9.28
lazy-object-proxy==1.4.3
macholib==1.16.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
pip-tools==7.4.1
platformdirs==4.3.6
propcache==0.2.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pylint==2.6.0
pyproject_hooks==1.2.0
pytz==2020.1
PyYAML==6.0.2
regex==2024.11.6
ruff==0.7.3
setuptools==75.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
tqdm==4.67.0
wheel==0.44.0
wrapt==1.12.1
yarl==1.15.1
